pyproject.toml
[project]
name = "noobrobot"
version = "0.1.0"
description = "Robot Framework MCP Tools for automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "robotframework>=6.1.1",
    "pydantic>=2.0.0",
    "starlette>=0.28.0",
    "uvicorn>=0.23.2",
    "mcp>=0.1.0", # Machine Control Protocol
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.4",
    "jinja2>=3.1.2",
    "selenium>=4.10.0", # Web automation
    "robotframework-seleniumlibrary>=6.1.0", # Robot Framework Selenium integration
    "webdriver-manager>=3.8.6", # For automatic ChromeDriver management
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
noobrobot = "main:main"
